requirements.txt
streamlit
pandas
plotly
fpdf
python-dotenv
numpy
scikit-learn